+++ prequ/writer.py
@@ -7,7 +7,8 @@ from ._compat import ExitStack
 from .file_replacer import FileReplacer
 from .logging import log
 from .utils import (
-    UNSAFE_PACKAGES, comment, dedup, format_requirement, key_from_req)
+    UNSAFE_PACKAGES, comment, dedup, format_requirement, formatted_as,
+    key_from_ireq)
 
 
 class OutputWriter(object):
@@ -31,7 +32,9 @@ class OutputWriter(object):
         self.silent = silent
 
     def _sort_key(self, ireq):
-        return key_from_req(ireq.req) if ireq.req else u'{}'.format(ireq)
+        line_format = formatted_as(ireq, self.find_links)
+        section_num = {'path': 0}.get(line_format, 9)
+        return (section_num, key_from_ireq(ireq))
 
     def write_header(self):
         if self.emit_header:
@@ -97,7 +100,7 @@ class OutputWriter(object):
         for ireq in packages:
             line = self._format_requirement(
                 ireq, reverse_dependencies, primary_packages,
-                markers.get(key_from_req(ireq.req)), hashes=hashes)
+                markers.get(key_from_ireq(ireq)), hashes=hashes)
             yield line
 
         if unsafe_requirements:
@@ -109,7 +112,7 @@ class OutputWriter(object):
                 req = self._format_requirement(ireq,
                                                reverse_dependencies,
                                                primary_packages,
-                                               marker=markers.get(key_from_req(ireq.req)),
+                                               marker=markers.get(key_from_ireq(ireq)),
                                                hashes=hashes)
                 if not allow_unsafe:
                     yield comment('# {}'.format(req))
@@ -132,14 +135,18 @@ class OutputWriter(object):
                     f.write(os.linesep.encode('utf-8'))
 
     def _format_requirement(self, ireq, reverse_dependencies, primary_packages, marker=None, hashes=None):
-        line = format_requirement(ireq, marker=marker)
+        line = format_requirement(
+            ireq,
+            marker=marker,
+            root_dir=os.path.dirname(self.dst_file),
+            find_links_dirs=self.find_links)
 
         ireq_hashes = (hashes if hashes is not None else {}).get(ireq)
         if ireq_hashes:
             for hash_ in sorted(ireq_hashes):
                 line += " \\\n    --hash={}".format(hash_)
 
-        if not self.annotate or key_from_req(ireq.req) in primary_packages:
+        if not self.annotate or key_from_ireq(ireq) in primary_packages:
             return line
 
         # Annotate what packages this package is required by