+++ tests/test_top_level_editable.py
@@ -7,13 +7,13 @@ from prequ.scripts._repo import get_pip_command
 
 
 class MockedPyPIRepository(PyPIRepository):
-    def get_dependencies(self, ireq):
+    def _get_dependencies(self, ireq):
         # "mock" everything but editable reqs to avoid disk and network I/O
         # when possible
         if not ireq.editable:
             return set()
 
-        return super(MockedPyPIRepository, self).get_dependencies(ireq)
+        return super(MockedPyPIRepository, self)._get_dependencies(ireq)
 
 
 def _get_repository():