+++ tests/test_utils.py
@@ -1,7 +1,21 @@
-from pytest import raises
+import os
+import shutil
+
+from pip.download import path_to_url
 
 from prequ.utils import (
-    as_tuple, dedup, flat_map, format_requirement, format_specifier)
+    as_tuple, dedup, flat_map, format_requirement, format_specifier,
+    is_subdirectory)
+
+
+def test_is_subdirectory():
+    cwd = os.getcwd()
+    test_dir = os.path.join(cwd, 'test')
+    assert is_subdirectory(cwd, test_dir)
+    assert is_subdirectory(os.path.join(test_dir, '..'), test_dir)
+    assert is_subdirectory(cwd, cwd)
+
+    assert not is_subdirectory(test_dir, cwd)
 
 
 def test_format_requirement(from_line):
@@ -12,6 +26,19 @@ def test_format_requirement(from_line):
 def test_format_requirement_editable(from_editable):
     ireq = from_editable('git+git://fake.org/x/y.git#egg=y')
     assert format_requirement(ireq) == '-e git+git://fake.org/x/y.git#egg=y'
+
+
+def test_format_requirement_non_relative_editable(from_editable, small_fake_package_dir, tmpdir):
+    tmp_package_dir = os.path.join(str(tmpdir), 'small_fake_package')
+    shutil.copytree(small_fake_package_dir, tmp_package_dir)
+    ireq = from_editable(tmp_package_dir)
+    assert format_requirement(ireq) == '-e ' + path_to_url(tmp_package_dir)
+
+
+def test_format_requirement_relative_editable(from_editable, small_fake_package_dir):
+    ireq = from_editable(small_fake_package_dir)
+    assert format_requirement(ireq, root_dir='.') == (
+        '-e ./tests/fake_pypi/small_fake_package')
 
 
 def test_format_specifier(from_line):
@@ -40,16 +67,15 @@ def test_as_tuple(from_line):
     assert version == '1.1'
     assert extras == ("extra1", "extra2")
 
-    # Non-pinned versions aren't accepted
-    should_be_rejected = [
+    # Non-pinned versions return None as version
+    non_pinneds = [
         'foo==1.*',
         'foo~=1.1,<1.5,>1.2',
         'foo',
     ]
-    for spec in should_be_rejected:
+    for spec in non_pinneds:
         ireq = from_line(spec)
-        with raises(TypeError):
-            as_tuple(ireq)
+        assert as_tuple(ireq)[1] is None
 
 
 def test_flat_map():